clean up test name, move getOrgOrigin to testUtils and construct url with it
